requirements.txt
Django==4.1
django-import-export==3.2.0
